+++ public/js/main.js
@@ -14,7 +14,26 @@ function initHideFlash() {
   }, 5000);
 }
 
+/*
+function initExpand () {
+  $( ".expand" ).each(function( index ) {
+  });
+}
+
+ */
+
+function displayDomainFields () {
+  var domainType = $('input[name="domain-type"]:checked').val();
+  $('.display-for-subdir, .display-for-domain, .display-for-subdomain').hide();
+
+  // console.log('.display-for'+domainType);
+  $('.display-for-'+domainType).show();
+}
+
 $(function () {
+
+
+  displayDomainFields ()
 
   // # CREATE SITE FORM VALIDATION
   var existingDomains = [];
@@ -31,13 +50,7 @@ $(function () {
   });
 
   $('.domain-type').change(function (event) {
-    var domainType = $('input[name="domain-type"]:checked').val();
-
-    if (domainType === 'domain') {
-      $('.wildcardHost').addClass('d-none');
-    } else {
-      $('.wildcardHost').removeClass('d-none');
-    }
+    displayDomainFields();
   });
 
   $('#create-site-form').validate({
@@ -58,8 +71,13 @@ $(function () {
   $(".valid-domain-character").keypress(function(event) {
       var key = event.which;
       var keychar = String.fromCharCode(key).toLowerCase();
+      var allowedCharacters = "abcdefghijklmnopqrstuvwxyz0123456789-.:";
 
-      if ((("abcdefghijklmnopqrstuvwxyz0123456789-.:/").indexOf(keychar) === -1)) {
+      if ($(this).hasClass('valid-domain-character-allow-slash')) {
+        allowedCharacters = allowedCharacters + '/';
+      }
+
+      if ((allowedCharacters).indexOf(keychar) === -1) {
          event.preventDefault();
          return false;
       }
@@ -71,9 +89,15 @@ $(function () {
   $(".valid-domain-character").on('input', function(event) {
     //also enfore lowercase
     var lowercaseValue = $(this).val().toLowerCase();
+    var regex = /[^a-zA-Z0-9-.:_]/g;
+
+
+    if ($(this).hasClass('valid-domain-character-allow-slash')) {
+      regex = /[^a-zA-Z0-9-.//:_]/g;
+    }
 
     // remove all chars that are not alpha numeric
-    lowercaseValue = lowercaseValue.replace(/[^a-zA-Z0-9-.://_]/g, "");
+    lowercaseValue = lowercaseValue.replace(regex, "");
 
     $(this).val(lowercaseValue);
   });