+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## unreleased
+* Fix site delete error in k8s
+
+## v0.21.0
+* Update react admin
+* Update proxy call to API
+* Add 'anonymize site' functionality
+* Add extraData to newsletter export
+
+## v0.20.0
+* Rewrite of ingress logic and adding support for subdir sites (domain.com/site1 &&domain.com/site2) for Kubernetes environments
+* For new sites set 2FA on by default for admin, editors & moderators
+
 ## v0.18.0
 * Fix shared images between copies of sites